+++ src/Screens/SettingsScreen/index.tsx
@@ -1,115 +1,33 @@
 
-import React, { useMemo } from 'react';
-import { View } from 'react-native';
+import React, { memo } from 'react';
 
-import { translations } from '@Translations/index';
-import { navigate } from '@Globals/NavigationControler';
-import ConfigService from '@Services/ConfigService';
-import DatabaseService from '@Services/DatabaseService';
-import CacheService from '@Services/CacheService';
-import ThemeService from '@Services/ThemeService';
-import HapticsService from '@Services/HapticsService';
-import AlertService from '@Services/AlertService';
+import { Loading } from '@Types/AppTypes';
 
-import { Button } from '@Button/index';
+import { Animation } from '@Animation/index';
 import { Layout } from '@Layout/index';
 import { TC } from './__TC__';
-import CredentialService from '@Services/CredentialService';
-import { Animation } from '@Animation/index';
+import { LC } from './__LC__';
 
-export default function SettingsScreen(): JSX.Element {
-
-  const config = useMemo(() => ConfigService.config, []);
-  const theme  = useMemo(() => ThemeService.appThemes[config.appTheme].component, []);
-  const R      = useMemo(() => translations.screen.settingsScreen[config.language], []);
+export const SettingsScreen = memo((props: {
+  settingsScopeState: Loading
+}) => {
 
   return (
     <Layout.Screen
       screenButtons={<TC.ScreenButtons />}
     >
-      <Animation.SlideFromLeft
-        delay={300}
-        duration={200}
-      >
-        <View
-          style={{ gap: 1 }}
+      {props.settingsScopeState === 'Loading' ? (
+        <Layout.Loading />
+      ) : (
+        <Animation.SlideFromLeft
+          delay={300}
+          duration={200}
         >
-          <Button.TextWithIcon
-            title={R['Language']}
-            iconName="language"
-            onPress={() => navigate('LANGUAGE SELECTION SCOPE')}
-            theme={{
-              font: theme.font_Button,
-              font_Pressed: theme.font_active,
-              background: theme.background_Button,
-              background_Pressed: theme.background_active,
-            }}
-          />
-          <Button.TextWithIcon
-            title={R['Themes']}
-            iconName="color-palette"
-            onPress={() => navigate('THEME SCOPE')}
-            theme={{
-              font: theme.font_Button,
-              font_Pressed: theme.font_active,
-              background: theme.background_Button,
-              background_Pressed: theme.background_active,
-            }}
-          />
-          <Button.TextWithIcon
-            title={R['Credentials']}
-            iconName="card-outline"
-            onPress={() => navigate('CREDENTIAL SCOPE')}
-            theme={{
-              font: theme.font_Button,
-              font_Pressed: theme.font_active,
-              background: theme.background_Button,
-              background_Pressed: theme.background_active,
-            }}
-          />
-          <Button.TextWithIcon
-            title={'Whipe All Data'}
-            iconName="trash-outline"
-            onPress={async () => await whipeAllData()}
-            theme={{
-              font: theme.background,
-              background: theme.wrong,
-              font_Pressed: theme.wrong,
-              background_Pressed: theme.background_active,
-            }}
-          />
-        </View>
-      </Animation.SlideFromLeft>
+          <Layout.ScrollView>
+            <LC.SettingsButtons />
+          </Layout.ScrollView>
+        </Animation.SlideFromLeft>
+      )}
     </Layout.Screen>
   );
-}
-
-async function whipeAllData() {
-  AlertService.handleAlert(true,
-    {
-      question: 'Want to whipe database?',
-      type: 'warning',
-    },
-    async () => {
-      HapticsService.vibrate('success');
-      await CredentialService.deleteCredentialsFolder();
-      await DatabaseService.deleteDatabaseFolder();
-      await CacheService.deleteLastOpenProject();
-      CacheService.lastOpenProject = {
-        id_project: '',
-        name: '',
-        rules: {},
-        sampleAlias: {
-          plural: '',
-          singular: '',
-        },
-      };
-      CacheService.allProjects = [];
-      CacheService.allWidgets_Project = [];
-      CacheService.allWidgets_Template = [];
-      CacheService.allSamples = [];
-      CacheService.allWidgets_Sample = [];
-      navigate('RESTART APP');
-    }
-  );
-}
+});