+++ src/Screens/TemplateScreen/__TC__/ScreenButtons.tsx
@@ -19,6 +19,19 @@ export const ScreenButtons = memo((props: {
   const projectSettings = useMemo(() => CacheService.getProjectFromCache(id_project), []);
 
   const onCreateWidget = useCallback(async () => {
+
+    // Project status update ============================================
+    const projectSettings = CacheService.getProjectFromCache(id_project);
+    if (projectSettings.status === 'uploaded') {
+      projectSettings.status = 'modified';
+      await ProjectService.updateProject(
+        projectSettings,
+        () => CacheService.updateCache_ProjectSettings(projectSettings),
+        (erroMessage) => alert(erroMessage)
+      );
+    }
+
+    // Widget creation ==============================
     const newWidget = ProjectService.getWidgetData();
     await ProjectService.createWidget_Template(
       id_project,
@@ -29,6 +42,7 @@ export const ScreenButtons = memo((props: {
       },
       (errorMessage) => alert(errorMessage)
     );
+
   }, [props.onWidgetCreation, id_project]);
 
   return (