+++ src/Services/ConfigService.ts
@@ -2,14 +2,18 @@ import { ConfigDTO } from '@Types/AppTypes';
 import LocalStorageService from './LocalStorageService';
 import LanguageService from './LanguageService';
 import ThemeService from './ThemeService';
+import DateTimeService from './DateTimeService';
 
 export default class ConfigService {
 
   static LOCAL_STORAGE_KEY: string = 'config';
+  static deviceLangauge = LanguageService.getDeviceLanguage();
   static config: ConfigDTO = {
-    language: LanguageService.getDeviceLanguage(),
-    appTheme: 'Dark',
-    widgetTheme: 'Light',
+    language:     this.deviceLangauge,
+    dateFormat:   DateTimeService.DateFormatByTag(this.deviceLangauge),
+    timeFormat:   DateTimeService.TimeFormatByTag(this.deviceLangauge),
+    appTheme:     'Dark',
+    widgetTheme:  'Light',
   };
 
   static async loadConfig(): Promise<void> {
@@ -28,10 +32,13 @@ export default class ConfigService {
   private static verifyConfigDTOIntegrity(dto: ConfigDTO): ConfigDTO {
 
     const { App, Widget } = ThemeService.themeNamesArray;
+    const deviceLanguage = LanguageService.getDeviceLanguage();
 
     const verifiedConfigDTO: ConfigDTO = {
-      language:    dto.language ?? LanguageService.getDeviceLanguage(),
-      appTheme:    App.includes(dto.appTheme)       ? dto.appTheme    : 'Dark',
+      language:    dto.language ?? deviceLanguage,
+      dateFormat:  dto.dateFormat ?? DateTimeService.DateFormatByTag(deviceLanguage),
+      timeFormat:  dto.timeFormat ?? DateTimeService.TimeFormatByTag(deviceLanguage),
+      appTheme:    App.includes(dto.appTheme) ? dto.appTheme : 'Dark',
       widgetTheme: Widget.includes(dto.widgetTheme) ? dto.widgetTheme : 'Light',
     };
     return verifiedConfigDTO;